+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Removed
 
+- イベント属性からaggregateEntranceGate,aggregateOfferを廃止
+
 ### Fixed
 
 ### Security